Enforce omeName use in LdapPasswordProvider.
This still fails with a "Can't find default group for" message.
It seems that after the LdapPasswordProvider call, SessionManagerImpl
still uses the original user-provider login string (in this case:
email address) as the value for the Principal and username.
Any query that assumes this string is the omeName will fail.